Applies first batch of requested changes
